+++ src/main/java/com/marginallyclever/ro3/AllPanels.java
@@ -1,6 +1,7 @@
 package com.marginallyclever.ro3;
 
 import com.formdev.flatlaf.FlatLightLaf;
+import com.marginallyclever.ro3.node.nodes.pose.poses.*;
 import com.marginallyclever.ro3.physics.ODEPhysicsPanel;
 import org.reflections.Reflections;
 import org.slf4j.Logger;
@@ -13,6 +14,10 @@ import java.util.Set;
 import javax.swing.*;
 import java.awt.*;
 
+/**
+ * This class should be in tests but cannot because {@link Reflections} does not work across the
+ * production/test boundary.
+ */
 public class AllPanels {
     private static final Logger logger = LoggerFactory.getLogger(AllPanels.class);
 
@@ -104,11 +109,11 @@ public class AllPanels {
             com.marginallyclever.ro3.node.nodes.marlinrobotarm.MarlinRobotArmPanel.class,
             ODEPhysicsPanel.class,
             com.marginallyclever.ro3.node.nodes.pose.PosePanel.class,
-            com.marginallyclever.ro3.node.nodes.pose.poses.AttachmentPointPanel.class,
-            com.marginallyclever.ro3.node.nodes.pose.poses.CameraPanel.class,
-            com.marginallyclever.ro3.node.nodes.pose.poses.LimbPanel.class,
-            com.marginallyclever.ro3.node.nodes.pose.poses.LookAtPanel.class,
-            com.marginallyclever.ro3.node.nodes.pose.poses.MeshInstancePanel.class,
+            AttachmentPointPanel.class,
+            CameraPanel.class,
+            LimbPanel.class,
+            LookAtPanel.class,
+            MeshInstancePanel.class,
             com.marginallyclever.ro3.texture.TextureChooserDialog.class
         ));
     }
