+++ src/main/java/com/marginallyclever/ro3/Registry.java
@@ -13,16 +13,16 @@ import com.marginallyclever.ro3.node.nodes.behavior.decorators.*;
 import com.marginallyclever.ro3.node.nodes.limbplanner.LimbPlanner;
 import com.marginallyclever.ro3.node.nodes.limbsolver.LimbSolver;
 import com.marginallyclever.ro3.node.nodes.marlinrobotarm.MarlinRobotArm;
-import com.marginallyclever.ro3.node.nodes.ode4j.CreatureController;
-import com.marginallyclever.ro3.node.nodes.ode4j.ODEHinge;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODEBox;
-import com.marginallyclever.ro3.node.nodes.ode4j.ODEPlane;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODECapsule;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODECylinder;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODESphere;
+import com.marginallyclever.ro3.node.nodes.odenode.CreatureController;
+import com.marginallyclever.ro3.node.nodes.odenode.ODEHinge;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODEBox;
+import com.marginallyclever.ro3.node.nodes.odenode.ODEPlane;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODECapsule;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODECylinder;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies.ODESphere;
+import com.marginallyclever.ro3.node.nodes.pose.poses.*;
 import com.marginallyclever.ro3.physics.ODEPhysics;
 import com.marginallyclever.ro3.node.nodes.pose.*;
-import com.marginallyclever.ro3.node.nodes.pose.poses.*;
 import com.marginallyclever.ro3.texture.TextureFactory;
 
 import javax.swing.event.EventListenerList;
