+++ src/main/java/com/marginallyclever/ro3/apps/commands/ImportMesh.java
@@ -1,18 +1,13 @@
 package com.marginallyclever.ro3.apps.commands;
 
 import com.marginallyclever.ro3.Registry;
-import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodes.pose.poses.MeshInstance;
-import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.swing.undo.AbstractUndoableEdit;
 import javax.swing.undo.CannotUndoException;
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.security.InvalidParameterException;
 
 /**
