+++ src/main/java/com/marginallyclever/ro3/node/nodes/odenode/CreatureController.java
@@ -1,8 +1,8 @@
-package com.marginallyclever.ro3.node.nodes.ode4j;
+package com.marginallyclever.ro3.node.nodes.odenode;
 
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.node.Node;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODEBody;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.ODEBody;
 import com.marginallyclever.ro3.physics.CollisionListener;
 import org.ode4j.ode.DContact;
 import org.ode4j.ode.DGeom;
