+++ src/main/java/com/marginallyclever/ro3/node/nodes/odenode/ODEHingePanel.java
@@ -1,15 +1,12 @@
-package com.marginallyclever.ro3.node.nodes.ode4j;
+package com.marginallyclever.ro3.node.nodes.odenode;
 
 import com.marginallyclever.ro3.PanelHelper;
 import com.marginallyclever.ro3.apps.nodeselector.NodeSelector;
 import com.marginallyclever.ro3.node.NodePath;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODEBody;
-import com.marginallyclever.ro3.node.nodes.ode4j.odebody.ODEBox;
-import com.marginallyclever.ro3.node.nodes.pose.Pose;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.ODEBody;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.ActionListener;
 import java.util.function.Consumer;
 
 /**
