+++ src/main/java/com/marginallyclever/ro3/node/nodes/odenode/odebody/odebodies/ODECapsule.java
@@ -1,8 +1,9 @@
-package com.marginallyclever.ro3.node.nodes.ode4j.odebody;
+package com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies;
 
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.mesh.shapes.Capsule;
 import com.marginallyclever.ro3.mesh.shapes.Sphere;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.ODEBody;
 import com.marginallyclever.ro3.physics.ODE4JHelper;
 import com.marginallyclever.ro3.node.nodes.pose.poses.MeshInstance;
 import org.json.JSONObject;
