+++ src/main/java/com/marginallyclever/ro3/node/nodes/odenode/odebody/odebodies/ODESphere.java
@@ -1,7 +1,8 @@
-package com.marginallyclever.ro3.node.nodes.ode4j.odebody;
+package com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies;
 
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.mesh.shapes.Sphere;
+import com.marginallyclever.ro3.node.nodes.odenode.odebody.ODEBody;
 import com.marginallyclever.ro3.physics.ODE4JHelper;
 import com.marginallyclever.ro3.node.nodes.pose.poses.MeshInstance;
 import org.json.JSONObject;
