+++ src/main/java/com/marginallyclever/ro3/node/nodes/odenode/odebody/odebodies/ODESpherePanel.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.ro3.node.nodes.ode4j.odebody;
+package com.marginallyclever.ro3.node.nodes.odenode.odebody.odebodies;
 
 import com.marginallyclever.convenience.swing.NumberFormatHelper;
 import com.marginallyclever.ro3.PanelHelper;
