+++ src/main/java/com/marginallyclever/ro3/physics/ODEPhysicsPanel.java
@@ -1,0 +1,11 @@
+package com.marginallyclever.ro3.physics;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class ODEPhysicsPanel extends JPanel {
+    public ODEPhysicsPanel() {
+        super();
+        // TODO add physics settings
+    }
+}
