+++ src/main/resources/com/marginallyclever/ro3/node/nodes/odenode/icons8-mechanics-16.png
