+++ src/test/java/com/marginallyclever/ro3/node/nodes/pose/poses/LimbTest.java
@@ -2,9 +2,7 @@ package com.marginallyclever.ro3.node.nodes.pose.poses;
 
 import com.marginallyclever.ro3.Registry;
 import com.marginallyclever.ro3.apps.actions.LoadScene;
-import com.marginallyclever.ro3.node.nodes.pose.poses.Limb;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
